Completed Product/product-service

+++ src/services/product-service/productService.js
@@ -1,7 +1,9 @@
 const axios = require("axios");
 
 exports.saveProduct = async (req) => {
-  const token = req.kauth.grant.access_token.token;// Extract authentication token from req
+  //const token = req.kauth.grant.access_token.token;// Extract authentication token from req
+  const token = req.cookies.jwt;
+  if (!token) throw new Error("Unauthorized");
   const url = `${process.env.API_GATEWAY_URL}/product-service/api/v1/product/save-product`;
 
   const response = await axios.post(url, req.body, {
@@ -40,33 +42,37 @@ exports.getProductById = async (req) => {
 };
 
 exports.updateProductDetails = async (req) => {
-  const token = req.kauth.grant.access_token.token;
-  const { productId } = req.query;
+  //const token = req.kauth.grant.access_token.token;
+  const token = req.cookies.jwt;
+  if (!token) throw new Error("Unauthorized");
+  const { id } = req.query;
   const url = `${process.env.PRODUCT_MS_URL}/product-service/api/v1/product/update-product-details`;
 
   const response = await axios.patch(url, req.body, {
     headers: { Authorization: `Bearer ${token}` },
-    params: { productId },
+    params: { id },
   });
 
   return response.data;
 };
 
 exports.deleteProductByID = async (req) => {
-  const token = req.kauth.grant.access_token.token;
-  const { productId } = req.query;
+  //const token = req.kauth.grant.access_token.token;
+  const token = req.cookies.jwt;
+  if (!token) throw new Error("Unauthorized");
+  const { id } = req.query;
   const url = `${process.env.PRODUCT_MS_URL}/product-service/api/v1/product/delete-product-by-id`;
 
   const response = await axios.delete(url, {
     headers: { Authorization: `Bearer ${token}` },
-    params: { productId },
+    params: { id },
   });
 
   return response.data;
 };
 
 exports.getProductByFiltering = async (req) => {
-  const token = req.kauth.grant.access_token.token;
+  //const token = req.kauth.grant.access_token.token;
   const {
     productName,
     minPrice,
